07/04

+++ Embedded/Robot_Raimbaud_Clement.X/asservissement.c
@@ -1,7 +1,7 @@
 #include "asservissement.h"
 #include "Robot.h"
 
-void SetupPidAsservissement(volatile PidCorrector* PidCorr, double Kp, double Ki, double Kd, double proportionelleMax, double integralMax, double deriveeMax)
+void SetupPidAsservissement(volatile PidCorrector* PidCorr, float Kp, float Ki, float Kd, float proportionelleMax, float integralMax, float deriveeMax)
 {
     PidCorr->Kp = Kp;
     PidCorr->erreurProportionelleMax = proportionelleMax; //On limite la correction due au Kp
